Auto-extract coffee intensities from Puerto Blest images using Claude Vision
Add a reusable vision helper (koffe/scrapers/vision.py) that sends coffee-card
images to Claude Haiku and parses acidity/body/sweetness bar charts. The Puerto
Blest scraper now collects all product images and sends the last one (the coffee
card) to the vision API, eliminating the need for manual intensity updates.

Safety: all failures (missing API key, download errors, bad JSON) gracefully
return None values, preserving existing DB data. Cost: ~$0.01 per scrape run.

+++ koffe/scrapers/base.py
@@ -19,9 +19,9 @@ class CoffeeData:
     origin_country: str | None = None
     process: str | None = None
     roast_level: str | None = None
-    acidity: int | None = None    # 1–5
-    sweetness: int | None = None  # 1–5
-    body: int | None = None       # 1–5
+    acidity: float | None = None    # 1–5
+    sweetness: float | None = None  # 1–5
+    body: float | None = None       # 1–5
     variety: str | None = None
     altitude_masl: int | None = None
     brew_methods: list[str] | None = None
